Move the core code in worker.ts to worker-runtime.ts so it can be imported from another module
